Rename freezr to index
